Optimize examples for better performance (#59)
* zeros -> emtpy in cuda

* disable cuda hint if is been built

* update examples with better perf

* bump version to 0.2.0

* update readme

* update index and readm

* clean up doc

+++ nerfacc/ray_marching.py
@@ -187,7 +187,7 @@ def ray_marching(
     if sigma_fn is not None:
         # Query sigma without gradients
         ray_indices = unpack_info(packed_info)
-        sigmas = sigma_fn(t_starts, t_ends, ray_indices)
+        sigmas = sigma_fn(t_starts, t_ends, ray_indices.long())
         assert (
             sigmas.shape == t_starts.shape
         ), "sigmas must have shape of (N, 1)! Got {}".format(sigmas.shape)
